Move ./run.py to the zulipterminal.cli module.
This makes it easier for setup.py to create a CLI tool from a method
inside of the package. It also required changing import paths for a few
modules, hence the touching of multiple files.
